Improve fork choice rule test coverage
