+++ test_libs/pyspec/eth2spec/test/helpers/state.py
@@ -1,4 +1,6 @@
-from eth2spec.test.helpers.block import sign_block
+from copy import deepcopy
+from eth2spec.test.helpers.attestations import get_valid_attestation
+from eth2spec.test.helpers.block import sign_block, build_empty_block_for_next_slot
 
 
 def get_balance(state, index):
@@ -36,3 +38,30 @@ def state_transition_and_sign_block(spec, state, block):
     spec.state_transition(state, block)
     block.state_root = state.hash_tree_root()
     sign_block(spec, state, block)
+
+
+def next_epoch_with_attestations(spec,
+                                 state,
+                                 fill_cur_epoch,
+                                 fill_prev_epoch):
+    assert state.slot % spec.SLOTS_PER_EPOCH == 0
+
+    post_state = deepcopy(state)
+    blocks = []
+    for _ in range(spec.SLOTS_PER_EPOCH):
+        block = build_empty_block_for_next_slot(spec, post_state)
+        if fill_cur_epoch and post_state.slot >= spec.MIN_ATTESTATION_INCLUSION_DELAY:
+            slot_to_attest = post_state.slot - spec.MIN_ATTESTATION_INCLUSION_DELAY + 1
+            if slot_to_attest >= spec.epoch_start_slot(spec.get_current_epoch(post_state)):
+                cur_attestation = get_valid_attestation(spec, post_state, slot_to_attest)
+                block.body.attestations.append(cur_attestation)
+
+        if fill_prev_epoch:
+            slot_to_attest = post_state.slot - spec.SLOTS_PER_EPOCH + 1
+            prev_attestation = get_valid_attestation(spec, post_state, slot_to_attest)
+            block.body.attestations.append(prev_attestation)
+
+        state_transition_and_sign_block(spec, post_state, block)
+        blocks.append(block)
+
+    return state, blocks, post_state
